Move users out of storage package
