+++ app/users/remote.go
@@ -1,0 +1,6 @@
+package users
+
+type RemoteUser struct {
+	ID uint
+	Email string
+}
