+++ cmd/db_migrate_down.go
@@ -1,7 +1,7 @@
 package cmd
 
 import (
-	db "github.com/lbryio/lbrytv/internal/storage"
+	"github.com/lbryio/lbrytv/internal/storage"
 
 	"github.com/spf13/cobra"
 )
@@ -14,6 +14,6 @@ var dbMigrateDown = &cobra.Command{
 	Use:   "db_migrate_down",
 	Short: "Unapply database migrations (rewind back to the initial state)",
 	Run: func(cmd *cobra.Command, args []string) {
-		db.Init().MigrateDown()
+		storage.Init().MigrateDown()
 	},
 }
