+++ cmd/db_migrate_up.go
@@ -1,7 +1,7 @@
 package cmd
 
 import (
-	db "github.com/lbryio/lbrytv/internal/storage"
+	"github.com/lbryio/lbrytv/internal/storage"
 
 	"github.com/spf13/cobra"
 )
@@ -14,6 +14,6 @@ var dbMigrateUp = &cobra.Command{
 	Use:   "db_migrate_up",
 	Short: "Apply unapplied database migrations",
 	Run: func(cmd *cobra.Command, args []string) {
-		db.Init().MigrateUp()
+		storage.Init().MigrateUp()
 	},
 }
