+++ internal/storage/conn.go
@@ -1,4 +1,4 @@
-package db
+package storage
 
 import (
 	"fmt"
